refactor(http): moved #getErrorMessage to `@marblejs/core` module
